+++ src/components/calcite-dropdown-group/calcite-dropdown-group.tsx
@@ -97,11 +97,16 @@ export class CalciteDropdownGroup {
   @Listen("calciteDropdownItemRegister") registerCalciteDropdownItem(
     event: CustomEvent<ItemRegistration>
   ) {
-    const item = {
-      item: event.target as HTMLCalciteDropdownItemElement,
+    const item = event.target as HTMLCalciteDropdownItemElement;
+
+    if (this.selectionMode === "none") {
+      item.active = false;
+    }
+
+    this.items.push({
+      item,
       position: event.detail.position,
-    };
-    this.items.push(item);
+    });
 
     event.stopPropagation();
   }
@@ -109,6 +114,11 @@ export class CalciteDropdownGroup {
   @Listen("calciteDropdownItemSelect") updateActiveItemOnChange(
     event: CustomEvent
   ) {
+    if (this.selectionMode === "none") {
+      event.stopPropagation();
+      return;
+    }
+
     this.requestedDropdownGroup = event.detail.requestedDropdownGroup;
     this.requestedDropdownItem = event.detail.requestedDropdownItem;
     this.calciteDropdownItemChange.emit({