+++ CHANGELOG.md
@@ -16,8 +16,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `calcite-dropdown-group` - `registerCalciteItemHasChanged` event has been renamed to `calciteDropdownItemChange`
 - `calcite-dropdown-group` - `registerCalciteDropdownGroup` event has been renamed to `calciteDropdownGroupRegister`
 - `calcite-dropdown-item` - `registerCalciteDropdownItem` event has been renamed to `calciteDropdownItemRegister`
-- `calcite-dropdown-item` - `calciteDropdownItemSelected` event has been renamed to `calciteDropdownItemSelect`
+- `calcite-dropdown-item` - `calciteDropdownItemSelected` event has been renamed to `calciteDropdownItemSelect` and is now internal.
 - `calcite-dropdown-item` - `closeCalciteDropdown` event has been renamed to `calciteDropdownClose`
+
+### Added
+
+- `calcite-dropdown` now has a read-only `selectedItems` prop that contains all selected items.
+- `calcite-dropdown` now emits `calciteDropdownSelect` when an item selection changes.
 
 ### Fixed